requirements.txt
annotated-types==0.7.0
APScheduler==3.11.0
certifi==2025.1.31
charset-normalizer==3.4.1
humanfriendly==10.0
idna==3.10
joblib==1.4.2
numpy==2.2.4
pandas==2.2.3
prometheus_client==0.21.1
pydantic==2.11.2
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
python_dateutil >= 2.8.2